% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_raw}
\alias{covid_raw}
\title{Raw data for COVID-19 in Mexico}
\format{
A dataframe with columns:
\describe{
\item{fecha_actualizacion}{date of update}
\item{origen}{if observation comes from USMER system or not}
\item{sector}{institution where patient was treated}
\item{entidad_um}{state of hospital were the patient was treated}
\item{sexo}{gender of the patient}
\item{entidad_nac}{state of birth}
\item{entidad_res}{state of residence}
\item{municipio_res}{town of residence}
\item{tipo_paciente}{if patient was hospitalized or not}
\item{fecha_ingreso}{date of hospitalization}
\item{fecha_sintomas}{date of first symptom}
\item{fecha_def}{date of death, if ocurred}
\item{intubado}{if intubation was needed}
\item{neumonia}{if patient presented pneumonia}
\item{edad}{age}
\item{nacionalidad}{if patient is mexican or not}
\item{embarazo}{if patient was pregnant}
\item{habla_langua_indi}{if patient speaks native language}
\item{diabetes}{if patient has diabetes}
\item{epoc}{if patient had COPD}
\item{amsa}{if patient has asthma}
\item{inmusupr}{if patient has immunosuppression}
\item{hipertension}{if patient has hipertension}
\item{otra_con}{if patient has other comorbidities}
\item{cardiovascular}{if patient has cardiovascular diseases}
\item{obesidad}{if patient has obesity}
\item{renal_cronica}{if patient has chronic kidney disease}
\item{tabaquismo}{if patient is a smoker}
\item{otro_caso}{if patient has contact with a covid positive}
\item{resultado}{result of covid19 test}
\item{migrante}{if patient is migran}
\item{pais_nacionalidad}{nacionality}
\item{pais_origen}{country the patient came from}
\item{uci}{if patient required intensive care}
}
}
\source{
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}
}
\usage{
covid_raw
}
\description{
Raw data for COVID-19 in Mexico
}
\examples{
\dontrun{
 covid_raw
}
}
\keyword{datasets}
