% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diccionarios}
\alias{diccionarios}
\title{Dictionaries to recode covid_raw data}
\format{
A list with the following data frames:
\describe{
\item{entidad}{dataframe with values and description for states variables}
\item{municipio}{dataframe with values and description for towns}
\item{nacionalidad}{dataframe with values and descripttion for nationality variables}
\item{origen}{dataframe with values and description for origen variable}
\item{paciente}{dataframe with values and description for tipo_paciente variable}
\item{resultado}{dataframe with values and description for resultado variable}
\item{sector}{dataframe with values and description for sector variable}
\item{sexo}{dataframe with values and description for sexo(gender) variable}
\item{sino}{dataframe with values and description for binary (yes/no) variables}
\item{guia}{description of variables in spanish}
}
}
\source{
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}
}
\usage{
diccionarios
}
\description{
Dictionaries to recode covid_raw data
}
\examples{
\dontrun{
 diccionarios$entidad

 diccionarios$origen

 diccionarios$resultado
}
}
\keyword{datasets}
